requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fake-useragent==1.5.1
filelock==3.14.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
flask-mongoengine==1.0.0
flask-paginate==2024.4.12
flask-cors==3.0.10
Flask-PyMongo==2.3.0
flask-talisman==1.1.0
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.5.0
google_search_results==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mkl==2021.4.0
mongoengine==0.28.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx>=2.5,<3.0
nltk==3.8.1
numpy==1.26.4
openai==0.28.0
outcome==1.3.0.post0
packaging==24.0
platformdirs==4.2.2
pluggy==1.5.0
preshed==3.0.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
PySocks==1.7.1
pytest==8.2.1
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.31.0
schedule==1.2.2
selenium==4.21.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tbb==2021.12.0
textblob==0.18.0.post0
thinc==8.2.3
torch==2.3.1
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
virtualenv==20.26.2
wasabi==1.1.2
weasel==0.3.4
webdriver-manager==4.0.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
gunicorn==20.1.0

